Introduce AbstractDirectionalBlock.
