+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/block/AbstractDirectionalBlock.java
@@ -1,0 +1,62 @@
+package com.refinedmods.refinedstorage2.platform.common.block;
+
+import com.refinedmods.refinedstorage2.platform.common.util.BiDirection;
+
+import javax.annotation.Nullable;
+
+import net.minecraft.core.Direction;
+import net.minecraft.world.item.context.BlockPlaceContext;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Rotation;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.StateDefinition;
+import net.minecraft.world.level.block.state.properties.EnumProperty;
+
+public abstract class AbstractDirectionalBlock extends AbstractBaseBlock {
+    private static final EnumProperty<BiDirection> DIRECTION = EnumProperty.create("direction", BiDirection.class);
+
+    protected AbstractDirectionalBlock(final Properties properties) {
+        super(properties);
+    }
+
+    protected BlockState getDefaultState() {
+        return super.getDefaultState().setValue(DIRECTION, BiDirection.NORTH);
+    }
+
+    @Override
+    protected void createBlockStateDefinition(final StateDefinition.Builder<Block, BlockState> builder) {
+        super.createBlockStateDefinition(builder);
+        builder.add(DIRECTION);
+    }
+
+    @Override
+    public BlockState getStateForPlacement(final BlockPlaceContext ctx) {
+        final BlockState state = defaultBlockState();
+        return state.setValue(
+            DIRECTION,
+            getDirection(ctx.getHorizontalDirection(), ctx.getPlayer() != null ? ctx.getPlayer().getXRot() : 0)
+        );
+    }
+
+    @Override
+    @SuppressWarnings("deprecation")
+    public BlockState rotate(final BlockState state, final Rotation rotation) {
+        final BiDirection currentDirection = state.getValue(DIRECTION);
+        return state.setValue(DIRECTION, currentDirection.rotate());
+    }
+
+    private BiDirection getDirection(final Direction playerFacing, final float playerPitch) {
+        if (playerPitch > 65) {
+            return BiDirection.forUp(playerFacing);
+        } else if (playerPitch < -65) {
+            return BiDirection.forDown(playerFacing.getOpposite());
+        } else {
+            return BiDirection.forHorizontal(playerFacing.getOpposite());
+        }
+    }
+
+    @Nullable
+    public static BiDirection getDirection(@Nullable final BlockState state) {
+        return state != null && state.hasProperty(DIRECTION) ? state.getValue(DIRECTION) : null;
+    }
+}
